# Remove json marshal call in common blocks.