In order to prevent memory allocations on creation of common blocks, this PR separates the caching layer (used in the harvester) and the common block entries.