@a-feld 2021-08-23
Before:
```
goos: darwin
goarch: amd64
pkg: github.com/newrelic/newrelic-telemetry-sdk-go/telemetry
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkLogCommonBlock-12               1482639               809.1 ns/op           648 B/op         12 allocs/op
BenchmarkMetricCommonBlock-12            1454265               824.8 ns/op           688 B/op         12 allocs/op
BenchmarkSpanCommonBlock-12              1473201               812.2 ns/op           648 B/op         12 allocs/op
PASS
```

After:
```
goos: darwin
goarch: amd64
pkg: github.com/newrelic/newrelic-telemetry-sdk-go/telemetry
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkLogCommonBlock-12               3110510               382.7 ns/op           376 B/op          5 allocs/op
BenchmarkMetricCommonBlock-12            3000152               400.1 ns/op           424 B/op          5 allocs/op
BenchmarkSpanCommonBlock-12              3112154               390.9 ns/op           376 B/op          5 allocs/op
PASS
```